+++ changelog.md
@@ -27,6 +27,7 @@
 - Added class `Checklists/Items`
 - Added class `ExternalLinkPatterns`
 - Added class `HttpApiModel`
+- Added class `ToDoItems`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`